+++ drizzle/0022_far_cannonball.sql
@@ -1,0 +1,18 @@
+CREATE TABLE `comment` (
+	`id` text PRIMARY KEY NOT NULL,
+	`account_id` text NOT NULL,
+	`song_id` text NOT NULL,
+	`user_id` text NOT NULL,
+	`title` text NOT NULL,
+	`body` text NOT NULL,
+	`at` real,
+	`edited_at` integer,
+	`created_at` integer DEFAULT (cast(unixepoch('subsecond') * 1000 as integer)) NOT NULL,
+	`updated_at` integer DEFAULT (cast(unixepoch('subsecond') * 1000 as integer)) NOT NULL,
+	FOREIGN KEY (`account_id`) REFERENCES `account`(`id`) ON UPDATE no action ON DELETE cascade,
+	FOREIGN KEY (`song_id`) REFERENCES `song`(`id`) ON UPDATE no action ON DELETE cascade,
+	FOREIGN KEY (`user_id`) REFERENCES `user`(`id`) ON UPDATE no action ON DELETE cascade
+);
+--> statement-breakpoint
+CREATE INDEX `comment_song_idx` ON `comment` (`song_id`,`created_at`);--> statement-breakpoint
+CREATE INDEX `comment_user_idx` ON `comment` (`user_id`);